feat(new-trace): File restructuring.
